TestSystem/software/py/requirements.txt
backcall>=0.1.0
certifi>=2018.10.15
cycler>=0.10.0
decorator>=4.3.0
fire>=0.1.3
h5py>=2.9.0
imageio>=2.4.1
ipython>=7.1.1
ipython-genutils>=0.2.0
jedi>=0.13.1
kiwisolver>=1.0.1
matplotlib>=3.0.2
numpy>=1.16.2
pandas>=0.24.2
tables>=3.5.1
parso>=0.3.1
pexpect>=4.6.0
pickleshare>=0.7.5
Pillow>=5.3.0
progressbar2>=3.38.0
prompt-toolkit>=2.0.7
ptyprocess>=0.6.0
Pygments>=2.2.0
pyparsing>=2.3.0
pyserial>=3.4
python-dateutil>=2.7.5
python-utils>=2.3.0
pyusb>=1.0.2
six>=1.11.0
timeout-decorator>=0.4.0
traitlets>=4.3.2
wcwidth>=0.1.7
pyyaml>=1.10.0
python-can>=2.2.0
jsonpickle>=1.0